realm: minor tweaks to appease MSVC

+++ test/realm/ctxswitch.cc
@@ -89,7 +89,7 @@ void sleep_task(const void *args, size_t arglen,
 }
 
 struct PriorityTestArgs {
-  long *counter;
+  int *counter;
   int exp_val1, exp_val2;
   int new_priority;
   Event wait_on;
@@ -119,7 +119,7 @@ void priority_child_task(const void *args, size_t arglen,
 void priority_master_task(const void *args, size_t arglen, 
 			  const void *userdata, size_t userlen, Processor p)
 {
-  long counter = 55;
+  int counter = 55;
 
   // We will create 5 workers that each try to increment the count twice.
   //  The first three are created before the master sleeps, and arrive at
